pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-observe"
version = "0.1.0"
description = "Framework-agnostic observability, audit, and eval for AI agent applications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Junjie Teoh", email = "junjieteoh@gmail.com" }
]
keywords = [
    "observability",
    "agents",
    "llm",
    "audit",
    "eval",
    "tracing",
    "enterprise",
    "guardrails",
    "safety",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
    "typing-extensions>=4.0;python_version<'3.11'",
]

[project.optional-dependencies]
viewer = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "jinja2>=3.1.0",
]
postgres = [
    "psycopg[pool]>=3.1.0",
]
otlp = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]
all = [
    "agent-observe[viewer,postgres,otlp,dev]",
]

[project.scripts]
agent-observe = "agent_observe.cli:main"

[project.urls]
Homepage = "https://github.com/junjieteoh/agent-observe"
Documentation = "https://github.com/junjieteoh/agent-observe#readme"
Repository = "https://github.com/junjieteoh/agent-observe"
Issues = "https://github.com/junjieteoh/agent-observe/issues"

[tool.hatch.build.targets.wheel]
packages = ["agent_observe"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["agent_observe"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "psycopg.*",
    "psycopg_pool.*",
    "opentelemetry.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (may require external services)",
]

[tool.coverage.run]
source = ["agent_observe"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
